requirements.txt
amqp==2.3.2
billiard==3.5.0.5
celery==4.2.1
certifi==2018.10.15
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.5.0
dj-static==0.0.6
Django==2.1.3
django-cors-headers==2.4.0
django-filter==2.0.0
django-filters==0.2.1
django-notifs==2.6.0
django-rest-swagger==2.2.0
djangorestframework==3.9.0
drf-nested-routers==0.91
flake8==3.6.0
gunicorn==19.9.0
idna==2.7
isort==4.3.4
itypes==1.1.0
Jinja2==2.10
kombu==4.2.2
MarkupSafe==1.1.0
mccabe==0.6.1
openapi-codec==1.3.2
pika==0.12.0
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
pycodestyle==2.4.0
pyflakes==2.0.0
pytz==2018.7
requests==2.20.1
simplejson==3.16.0
six==1.11.0
static3==0.7.0
uritemplate==3.0.0
urllib3==1.24.1
uWSGI==2.0.17.1
vine==1.1.4
votes==1.0.1
